Dashboard/requirements.txt
flask
flask-login
pymongo
python-dotenv
dnspython
chart.js